chore: organised layout of directory
